re-add  removed functionality
after the latest changes in DoorLock and Alarm area the widget is not
working as expected. Fixed the widget.

One item that still needs to be handled is that the Alarms - after
re-implemented don't send a notification to update the status of the
widget.

Example:
- Secure the door from widget
- unlock the door from the doorlock

the widget would still show that the door is locked.

I'm not sure the best way to handle this without causing other issues
and doing it in a generic way