Update homegenie.webapp.utility.js
The temperature that is displayed in the Groups drop down (i.e. Dashboard) and in the upper right corner of the Dashboard if it contains the Weather Widget with a decimal place and one significant digit.

I made some changes to the Weather Widget to round the temperature up on that (along with the Fahrenheit display changes in another pull request) and I was able to find that making these changes as noted will round the temp display on the main Dashboard and Group drop down.  Not sure if this would be better as an option or just allow the no decimal place code to be the default but wanted to share what I found.

Thanks