chore(licensing): Relicense project from GPL-3.0 to Apache 2.0
The entire HomeGenie codebase has been relicensed from the GNU GPLv3 to the Apache License, Version 2.0.

This change is intended to encourage broader adoption and to allow for easier integration with other open-source and commercial projects, aligning HomeGenie with the licensing standards of the modern .NET ecosystem.

This commit also removes obsolete Windows setup files from the assets folder.